modify fermi dirac distribution
